Migrate JNP batching internals to JAX compat
