+++ jax2onnx/plugins/jax/_batching_utils.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Sequence, Callable, cast
+from typing import Any, Sequence
 
 import numpy as np
 from jax import lax
@@ -10,44 +10,11 @@ from jax import lax
 from jax2onnx._compat.jax import (
     NOT_MAPPED,
     batching,
+    definitely_equal_shape,
     ensure_batching_not_mapped_attr,
 )
 
 ensure_batching_not_mapped_attr()
-
-
-def _resolve_definitely_equal_shape() -> Callable[[Any, Any], bool]:
-    try:  # Prefer the internal helper when available (moved in newer JAX versions).
-        from jax._src import core as _core_src
-
-        return cast(Callable[[Any, Any], bool], _core_src.definitely_equal_shape)
-    except Exception:  # pragma: no cover - fallback for older/older-stub JAX
-        try:
-            from jax import core as _core_public
-
-            return cast(
-                Callable[[Any, Any], bool],
-                getattr(_core_public, "definitely_equal_shape"),
-            )
-        except Exception:  # pragma: no cover - minimal fallback
-
-            def _fallback(s1: Any, s2: Any) -> bool:
-                if len(s1) != len(s2):
-                    return False
-                for d1, d2 in zip(s1, s2):
-                    if d1 is d2:
-                        continue
-                    try:
-                        if d1 != d2:
-                            return False
-                    except Exception:
-                        return False
-                return True
-
-            return _fallback
-
-
-_definitely_equal_shape: Callable[[Any, Any], bool] = _resolve_definitely_equal_shape()
 
 
 def _handle_scalar_broadcasting(ndim: int, x: Any, dim: Any) -> Any:
@@ -65,7 +32,7 @@ def broadcast_batcher_compat(
 
     shape, dim = next((x.shape, d) for x, d in zip(args, dims) if d is not NOT_MAPPED)
     if all(
-        _definitely_equal_shape(shape, x.shape) and d == dim
+        definitely_equal_shape(shape, x.shape) and d == dim
         for x, d in zip(args, dims)
         if np.ndim(x)
     ):
